refactor: 일정 삭제 시 작성자 검증 로직 추가
- 기존: 로그인 여부와 관계없이 일정 삭제 가능
- 변경: 로그인된 사용자와 일정 작성자의 정보를 비교하여 삭제 가능하도록 수정
- 예외 처리 추가: 작성자가 아닌 사용자가 삭제 시도 시 권한 없음(FORBIDDEN) 오류 반환

+++ src/main/java/com/example/schedulehub/service/ScheduleServiceImpl.java
@@ -76,10 +76,23 @@ public class ScheduleServiceImpl implements ScheduleService{
 
     // 인증 필요
     @Override
-    public void deleteSchedule(Long id) {
+    public void deleteSchedule(Long scheduleId, Long userId) {
 
-        Schedule findSchedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
+        if(isPasswordCorrect(scheduleId, userId)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
+        Schedule findSchedule = scheduleRepository.findScheduleByIdOrElseThrow(scheduleId);
 
         scheduleRepository.delete(findSchedule);
+    }
+
+    private boolean isPasswordCorrect(Long scheduleId, Long userId){
+
+        String creatorPassword = scheduleRepository.findScheduleByIdOrElseThrow(scheduleId).getUser().getPassword();
+
+        String accessPassword = userRepository.findUserByIdOrElseThrow(userId).getPassword();
+
+        return creatorPassword.equals(accessPassword);
     }
 }
